refactor(frontend): Add course module
