+++ frontend/src/app/assignment/assignment-routing.module.ts
@@ -1,10 +1,6 @@
 import {NgModule} from '@angular/core';
 import {RouterModule, Routes} from '@angular/router';
-
-import {CourseComponent} from './pages/course/course.component';
-import {CreateCourseComponent} from './pages/create-course/create-course.component';
 import {CreateSolutionComponent} from './pages/create-solution/create-solution.component';
-import {MyCoursesComponent} from './pages/my-courses/my-courses.component';
 import {MySolutionsComponent} from './pages/my-solutions/my-solutions.component';
 import {OverviewComponent} from './pages/overview/overview.component';
 import {SettingsComponent} from './pages/settings/settings.component';
@@ -13,10 +9,10 @@ const routes: Routes = [
   {path: '', component: OverviewComponent, data: {title: 'Assignments'}},
   {path: 'solutions', component: MySolutionsComponent, data: {title: 'My Solutions'}},
   {path: 'settings', component: SettingsComponent, data: {title: 'Settings'}},
-  {path: 'courses/create', component: CreateCourseComponent, data: {title: 'Create Course'}},
-  {path: 'courses/:cid', component: CourseComponent, data: {title: 'Course'}},
-  {path: 'courses/:cid/assignments/:aid', component: CourseComponent, data: {title: 'Course'}},
-  {path: 'courses', component: MyCoursesComponent, data: {title: 'My Courses'}},
+  {
+    path: 'courses',
+    loadChildren: () => import('./modules/course/course.module').then(m => m.CourseModule),
+  },
   {
     path: 'create',
     loadChildren: () => import('./modules/edit-assignment/edit-assignment.module').then(m => m.EditAssignmentModule),
