+++ frontend/src/app/assignment/assignment.module.ts
@@ -11,25 +11,18 @@ import {DndModule} from 'ngx-drag-drop';
 import {SharedModule} from '../shared/shared.module';
 import {AssignmentRoutingModule} from './assignment-routing.module';
 import {AssignmentSharedModule} from './modules/shared/shared.module';
-
-import {CourseComponent} from './pages/course/course.component';
-import {CreateCourseComponent} from './pages/create-course/create-course.component';
 import {CreateSolutionComponent} from './pages/create-solution/create-solution.component';
 import {MyAssignmentsComponent} from './pages/my-assignments/my-assignments.component';
-import {MyCoursesComponent} from './pages/my-courses/my-courses.component';
 import {MySolutionsComponent} from './pages/my-solutions/my-solutions.component';
 import {OverviewComponent} from './pages/overview/overview.component';
+import {SettingsComponent} from './pages/settings/settings.component';
 import {TokenModalComponent} from './pages/token-modal/token-modal.component';
 import {ConfigService} from './services/config.service';
-import { SettingsComponent } from './pages/settings/settings.component';
 
 @NgModule({
   declarations: [
-    CourseComponent,
-    CreateCourseComponent,
     CreateSolutionComponent,
     MyAssignmentsComponent,
-    MyCoursesComponent,
     MySolutionsComponent,
     TokenModalComponent,
     OverviewComponent,
