+++ frontend/src/app/assignment/modules/course/course.module.ts
@@ -1,0 +1,35 @@
+import {CommonModule} from '@angular/common';
+import {NgModule} from '@angular/core';
+import {FormsModule} from '@angular/forms';
+import {NgbTooltipModule, NgbTypeaheadModule} from '@ng-bootstrap/ng-bootstrap';
+import {ClipboardModule} from 'ngx-clipboard';
+import {DndModule} from 'ngx-drag-drop';
+import {SharedModule} from '../../../shared/shared.module';
+import {AssignmentSharedModule} from '../shared/shared.module';
+
+import {CourseRoutingModule} from './course-routing.module';
+import {CourseComponent} from './course/course.component';
+import {CreateCourseComponent} from './create-course/create-course.component';
+import {MyCoursesComponent} from './my-courses/my-courses.component';
+
+
+@NgModule({
+  declarations: [
+    CourseComponent,
+    CreateCourseComponent,
+    MyCoursesComponent,
+  ],
+  imports: [
+    CommonModule,
+    CourseRoutingModule,
+    SharedModule,
+    AssignmentSharedModule,
+    FormsModule,
+    NgbTypeaheadModule,
+    DndModule,
+    ClipboardModule,
+    NgbTooltipModule,
+  ],
+})
+export class CourseModule {
+}
