+++ frontend/src/app/assignment/modules/course/course/course.component.html
@@ -22,6 +22,6 @@
     </div>
   </div>
   <div class="col-md-9">
-    <app-create-solution *ngIf="assignmentID"></app-create-solution>
+    <!-- TODO <app-create-solution *ngIf="assignmentID"></app-create-solution> -->
   </div>
 </div>
